fix: update fee rate input handling to use parseFloat for better precision
- Changed input parsing from parseInt to parseFloat in FeeSettingsScreen and SendScreen to allow for decimal values in fee rate settings.
- Enhanced logging for custom fee rate changes to improve user feedback and debugging.

+++ app/(tabs)/send.tsx
@@ -119,7 +119,7 @@ export default function SendScreen() {
   const handleCustomFeeRateChange = (rate: string) => {
     console.log(`🔧 Send screen: Custom fee rate changed to ${rate}`);
     
-    const numericValue = parseInt(rate) || 0;
+    const numericValue = parseFloat(rate) || 0;
     
     // Update wallet store directly
     const updatedSettings = {
